refactor: createOrganization 위치 변경
