+++ client/src/api/mutations/useCreateOrganization.ts
@@ -1,10 +1,33 @@
 import { useMutation } from '@tanstack/react-query';
 
-import { createOrganization } from '../queries/organization';
+import { fetcher } from '../fetcher';
 import {
   CreateOrganizationAPIRequest,
   CreateOrganizationAPIResponse,
 } from '../types/organizations';
+
+export async function createOrganization(
+  body: CreateOrganizationAPIRequest
+): Promise<CreateOrganizationAPIResponse> {
+  const formData = new FormData();
+
+  const organization = {
+    name: body.organization.name.trim(),
+    description: body.organization.description.trim(),
+    nickname: body.organization.nickname.trim(),
+  };
+
+  formData.append(
+    'organization',
+    new Blob([JSON.stringify(organization)], { type: 'application/json' })
+  );
+
+  if (body.thumbnail) {
+    formData.append('thumbnail', body.thumbnail);
+  }
+
+  return fetcher.post<CreateOrganizationAPIResponse>('organizations', formData);
+}
 
 export const useCreateOrganization = () =>
   useMutation<CreateOrganizationAPIResponse, Error, CreateOrganizationAPIRequest>({
